Use expected type to guide type inference
Summary:
This is the first part of a new approach to type inference, making use of the "expected type" of an expression
in order to guide type inference. There are a number of situations in which lambdas aren't getting checked at
all, because the type information isn't available at the right moment. For example, consider this test:
```
class MyClosureRunner<-T> {
  public function __construct(private (function(T): string) $closure) {}
}

function expectsMyClosureRunnerMyClass(
  MyClosureRunner<MyClass> $x,
): MyClosureRunner<MyClass> {
  return $x;
}
function getThing1(): MyClosureRunner<MyClass> {
  $var = expectsMyClosureRunnerMyClass(
    new MyClosureRunner(
      function($arg) {
        $arg->missing_method();
        return "ok";
      },
    ),
  );
  return $var;
}
```
The body of the lambda is checked under the assumption that `$arg` has type `Tany`, which means that the errors
in the body are ignored.

The new approach is to do "bidirectional" type checking, where the inference function takes an optional type
argument that is the type that is "expected". Often this is known: in particular, when calling a function (e.g. `expectsMyClosureRunnerMyClass` above),
or in a `new` expression (e.g. `new MyClosureRunner` above). Eventually this information will be used to pass type information
through to the parameters (e.g. `$arg` above) so that the lambda can be checked at the point of use. For now, we pass
information around and call `check_expected_ty` whenever we have the expected type and inferred type in our hand. Even this
catches more errors, as the tests demonstrate.

The feature is currently under an experimental flag, enabled by default in `hh_single_type_check` but requiring the following for `hh_client` in the `.hhconfig` file:
```
enable_experimental_tc_features = contextual_inference
```
Other changes: the special typing of `array_map` wasn't flexible enough in its return type. Some logging improvements.

Also, we optionally log the number of lambdas that are checked (a) with fully declared parameters, (b) with contextual inference, and (c) using the legacy approach. Set `sharedmem_log_level=1` in the `.hhconfig` file and information will be placed in the log file.

TODO: we are reporting some lambdas to be recursive when they aren't. This seems to happen when there is a lambda without contextual type information nested in the body of a lambda *with* contextual type information. This will be fixed in a future diff.
TODO: avoid multiple checks. Currently we end up checking subtypes multiple times: the "expected type" parameter to the `expr` function can be ignored, and the caller doesn't get to know this. In future, we will probably pass the burden of checking that the expression actually has the expected type onto the callee.

Reviewed By: jamesjwu

Differential Revision: D5814170

fbshipit-source-id: 3ec602b55cf6075cdce0f105b760c0c39a25b6d0