Fix HHBBC bug when merging equivalent locals state
Summary:
In HHBBC, when merging equivalent locals state, we were calling
killLocEquiv() with the current local's next equivalent local, not the local
itself. Since the equivalencies are circular this worked in most cases, but its
possible for this to cause equivalencies which aren't true to be
propagated. This can cause HHBBC to derive incorrect facts and optimize
incorrectly.

Reviewed By: markw65

Differential Revision: D5900562

fbshipit-source-id: 04e97d2788a8b72a9ff25bce5865dbd73a84ca04