Fix smashable instructions in PPC64
Summary:
Currently on PPC64 architecture all immediate are loaded using up to 5
instructions, which makes the read not atomic.
The smashable instructions may have its immediate changed on the fly, i.e.,
while running or re-emitting the JITted code, which causes race condition,
if reading the immediate is composed with more than one instruction.
The race condition happen at two different moments:

1. The JITted code is being executed and the pipeline stop reading the
instructions group that reads the immediate in the middle (thread
preemption), then another thread smashes the instruction that uses the same
immediate.

2. The Decoder is reading the immediate used in a smashable instruction and
this thread stop being executed (thread preemption), then another thread
smashes the instruction that uses the same immediate.

For the reason above, this PR is being sent to enable the concept of TOC
(Table Of Contents) in the JITted code, where all immediate are save in
the "TOC" area, keeping the entire value in a slot of 64 bits, making the
read/write atomic, as only one operation is needed.
The TOC memory is an vector, where the pointer to it will be placed in the
R2 register, using up to two instructions to calculate its offset, which
keeps fixed, as only the position is TOC memory is changes when smashing.
Closes https://github.com/facebook/hhvm/pull/7959

Differential Revision: D5727325

Pulled By: mxw

fbshipit-source-id: d383b064e49b72e703aedf72f0dedace02a1d408