Add fanout to extends dependencies
Summary:
See T21568106 for why this change is necessary.

This diff adds the phase 2 fanout that we need in order to correctly initialize from saved state. I can guarantee there will be a regression on initialization from this change, but it's the safest thing to do for now.

I considered digging deep into the declaration code to try to do what dlreeves suggested and be smarter about adding dependencies, but this seems unnecessary given all the changes we'll be doing with incremental init that will improve this once again. This particular code path should be pretty much unused if we can get incremental init working, and after that it'll be more about improving our incremental mode, not this one use case.

Reviewed By: alexchow

Differential Revision: D5876155

fbshipit-source-id: 7d7c1380bbb40a1f98da76f97e736378b7890b3b