Introduce new dependency service
Summary:
This is a bit early, but I wanted to try the dependency analyzer out on www so I built out the new arguments necessary to test and create saved states with the new dependency graph instead of the current one.

This diff adds a new argument to hh_server --use-gen-deps, which, if passed with --save-mini, will use the new gen deps to save a saved state instead of the standard one. This new initialization path does not typecheck and does not go through the normal initialization in serverInit, avoiding all unnecessary work except those that will result in the dependency graph and fileInfo objects, which is what we need for saving a saved state.

Reviewed By: dabek

Differential Revision: D5752028

fbshipit-source-id: 70fa3ae403eee2ea37d3122c8caee64a4f77982c