Only use from_source range-binding when part of a chunk group is outside the range
Summary:
In D5853970, we started binding splits outside the formatting range to their solution in the original source. Unfortunately, the predicate we used ensures that we go through this code path for *every* chunk group, not just the ones that have splits that fall outside the range boundaries.

This change ensures that we only do this binding work when the range start or end falls inside a chunk group, instead of every time the range overlaps with a chunk group.

Differential Revision: D5908938

fbshipit-source-id: 5b25df46d64f8d9164bbdc02acbc3f65f47775a0