Ensure container literals are governed by only one rule
Summary:
Due to an oversight, the splits separating items in container literals can be governed by a different rule that the splits before the first item and after the last item. If the former breaks but not the latter, we get poor formatting like this:

```
$v = vec['one',
  'two'];
```

Until recently, this problem was masked by the fact that the rule governing the outer splits was Parental--if the inner splits broke, the rule governing the outer splits would break too. We don't enforce these relationships as strongly in the new range-binding codepath (intentionally, since code outside the range may be formatted in some style that hackfmt refuses to produce), so the mismatched rules in this case became visible.

Differential Revision: D5908963

fbshipit-source-id: 3247203a51972d8b1fd43c55f63d6e344dc616be