Track the source location of trailing commas
Summary:
Previously, we never tracked the location of trailing commas in lists (such as function arguments and container literals), since they may be deleted or inserted by hackfmt, depending on whether the list was collapsed onto one line or broken across several.

Instead, we would simply print the trailing comma if and only if the atom preceding it was printed. This generally produced the correct result, with one exception: when the range ended immediately before a trailing comma in the original source, we'd end up with two commas in the output.

With this change, we carry the source location of trailing commas which were present in the input all the way through hackfmt. This means that we can correctly choose to include or exclude trailing commas which fall at the beginning or end of a range.

Differential Revision: D5908984

fbshipit-source-id: a887e0e04e1e2cf2f84303ca9cbfbde440d06533