add tuple support for constants
Summary:
D4990753 and D5082235 added support for all the good stuff like vec/keyset/dict, arrays, and shapes. We missed `tuple`, because it's parsed as a function call and eventually dealt with here: diffusion/FBS/browse/master/fbcode/hphp/hack/src/naming/naming.ml$2147,2203

This, however, happens *after* constant checking. We deal with the raw AST, so we need to figure it out as a function call.

Reviewed By: dlreeves

Differential Revision: D5913209

fbshipit-source-id: 4223d8a1cf6322173697ad362dd396fb82ee12c9