New hg command for querying old files
Summary:
This is the first in a series of diffs to improve our incremental mode behavior.

When starting from saved state, we normally load the fileinfo object, parse all dirty files and typecheck all the dirty file's dependencies. This means that if you make a tiny change to one file and your one file happens to have a lot of dependencies, even if this change was minimal, we end up taking forever.

This is a series of diffs to do the following instead:

1. Load from saved state
2. Grab the saved state's base revision.
3. Query hg for the dirty files *old* versions, aka the versions that were present when the saved state was saved.
4. Parse and declare the dirty files.
5. Run incremental mode with dirty files as "need to be reparsed"

In my experiments, adding a new line to FBType.php currently takes 200 seconds on my devserver. It takes 20 seconds with the later changes, so I think this is worth doing.

Reviewed By: alexchow

Differential Revision: D5847169

fbshipit-source-id: c5437c813be9eced88086a066811e7314e29fa2c