New relative path for temp directory files
Summary: hg cat only allows us to save files to a directory, so we'll be saving the old files to a tmp directory inside /tmp/hh_server, and immediately deleting them after parsing it. This is the most error free way to grab the hg info. Otherwise, hg cat would return all the files in one big string, which is very difficult to parse as separate files.

Reviewed By: alexchow

Differential Revision: D5847171

fbshipit-source-id: 07ffaf8508660c6390a40720e2b144c80bd327b9