Convert PopL/PushL pairs to UnsetL
Summary:
This sequence takes a value off the stack, stores it in a local, then
copies it back onto the stack and unsets the local; we might as well
just unset it. In almost all cases, the unset is dead, or the local is
already uninit, and the Unset is optimized away.

Reviewed By: ricklavoie

Differential Revision: D5913152

fbshipit-source-id: d9fc3931eb66f5b5da52b7e4bf7d23a3fa9b4c62