Fix retranslate all feature for PPC64
Summary:
The concept of resizable buffers was not added to PPC64 specific files,
which is used for the retranslate all feature.
The decoder was changed to receive the destination and buffer addresses,
and the calls for it, in the relocation method, was changed to add the
buffer address.
Also methods to obtain the buffer address, based on the destination code
address, were added.
Closes https://github.com/facebook/hhvm/pull/7978

Reviewed By: paulbiss

Differential Revision: D5794190

Pulled By: mofarrell

fbshipit-source-id: a0861efba6faa30ac9ae10515b7c3a0469f17844