Parse old versions of dirty files in hg
Summary:
This diff introduces a new type of initialization, deemed "incremental init", mainly because it utilizes our incremental mode.

It utilizes the mercurial commands we introduced in the last few diffs in order to initialize and load into a state equivalent to when we saved our saved state. It then runs incremental mode on the dirty files to obtain the final error list.

I will add some secondary behavior in a later diff which checks for the number of dirty files and decides not to do incremental init if the list of files is too big.

Reviewed By: alexchow

Differential Revision: D5847172

fbshipit-source-id: 28ee3ec6ac54e89c4bf852fdd81e7fa32aa5cff6