Fix crash from passing bad object callable to array_map and family
Summary:
When lowering DecodeCufIter to IR, we were popping from the stack before
potentially side-exiting. So, if we side-exit because of a bad object callable,
the stack will have been popped when it shouldn't be. This causes problems when
we try to resume from the InterpOne. Defer the popping of the stack until after
we might side-exit.

Also fix a typo which was causing us to print the immediates of some bytecodes
improperly.

Reviewed By: mxw

Differential Revision: D5923284

fbshipit-source-id: e3d95e386aa2fb658520a169365978ac9b06f5a2