Emit Self and Parent for static properties, and FPushClsMethod[F]
Summary:
We did this originally because before hhir, we couldn't optimize
Self/Parent properly in the jit. But now we can, and they produce
smaller bytecode, and are faster in the interpreter (they should also
be marginally faster to jit, because they don't involve looking up a
class, and produce exactly the same code).

Reviewed By: ricklavoie

Differential Revision: D5911787

fbshipit-source-id: 12b656576cc17462a3a75464fc9d2b59d4eca06d