enable source locations
Summary:
This diff enables source locations for both HHVM and HackC. At this point semdiff ignores `.srcloc` directives so results of tests are mainly untouched however later we'll add validation for function spans and emitted srcloc directives.

Changes in semdiff are necessary to handle issue that were hidden before: semdiff crashed when parsing .hhas, then we resort to case-insensitive syntax diff and it showed that files are identical. Now when we emit source locations, semdiff still used to crash but now syntax diff did not consider files identical because of differences in source locations.

Reviewed By: aaronkroth

Differential Revision: D5921939

fbshipit-source-id: b5fd122d89adb584f64467d918b5d7c906a634fb