Get right brace token without trailing whitespace
Summary:
Motivation:
It looks like whitespace after the right brace in an interpolated string of the form `"${variable}"` is being trimmed.

e.g.

`" in ${a} on line ${b}"`
would produce hhas similiar to:
`String "on line "`
instead of
`String " on line "`

Fix:
The fix here is to use `next_token_no_trailing` when getting the right brace token.

Reviewed By: aaronkroth

Differential Revision: D5932022

fbshipit-source-id: 2a383fdc0db4a3d33a6d779547bfeb64d2f21930