Use expected type when checking collection/shape expressions
Summary:
Extending the work of D5814170 to pass contextual type information through when checking collection
construction, tuple construction, pair construction, and shape construction. For example, if the
type-checker expects type `(int,(function(C):bool))` when checking the expression `tuple($x,$y ==> $y->foo())`, then it passes the types `int` and `function(C):bool` down to the checking of the components of the pair.

Reviewed By: jamesjwu

Differential Revision: D5920661

fbshipit-source-id: 22420f469a13ec93bbf6a2b0824cec8912b2461b