Optimize InitThisLoc away
Summary:
hphpc uses heuristics to decide whether to use InitThisLoc/$this, or
just BareThis. The heuristics are based on analysis that I'd prefer to
leave until hhbbc time - so this makes hhbbc understand the behavior.

This can actually produce better results, because it will replace uses
of the thisLocal with BareThis anywhere they're known to have the same
value; whereas hphpc was all or nothing - if we needed an InitThisLoc,
we didn't use BareThis anywhere.

Reviewed By: ricklavoie

Differential Revision: D5937822

fbshipit-source-id: 332fce68d513a4b71b61ea317aa95281b54d8ab0