Fix an incorrect assumption in union_of
Summary:
union_of starts by checking if one Type is a subtype of the other. It
then asserts in DualDispatchUnion::operator()(SArray,SArray) that the
two arrays are not equal; but a could be TSArr and b could be
TOptSArrN, for example, so neither is a subtype of the other, but
their data might be.

Reviewed By: ricklavoie

Differential Revision: D5947066

fbshipit-source-id: 52f9aa77d7b39ee8d771437bf643323d3db2afa6