Fix a couple of ClsRef related dce bugs
Summary:
I noticed we don't even try to dce ClsRefGet* for objects, but when I
tried to fix it, I hit a crash with the new test case.

There were two issues:

First, we didn't properly handle the interaction between
forcedLiveLocations and slotLive; this code was largely duplicated
from the corresponding stack code, which doesn't have a corresponding
liveness set.

Second, we shouldn't have been setting forcedLiveLocations at all in
the test case - but we weren't clearing out the locations block-id
when we killed the slot UseInfo.

Reviewed By: ricklavoie

Differential Revision: D5950858

fbshipit-source-id: 5a018e3bdfe08cbdb045fd30ba41e562d498316d