hh rage
Summary:
If a user files a bug from the editor, say, I'd like to gather extra detailed information from hh_server. That way we can have richer bug reports. I've implemented this in a "rage" command.

The advantage of this over just writing to the server logs is that it's "on-demand", which means we can do computationally expensive work or be extra verbose in what we generate.

The difference between this and the existing "record" mechanism is that this doesn't require the user to know in advance that they're going to run into an issue. They can simply click the "file-a-bug" button and gather whatever information is available retrospectively.

In this diff, I've only made the rage command accessible to LSP clients -- not via the command line. And I proposed an extension to the LSP protocol for it. I trust that the LSP client has some sort of mechanism for automatically filing bugs based on what it gets. (We could decide to make a command-line "hh rage" command, but I'm not sure whether that would be practically used nor what information should go in it, so I left it off for now).

Reviewed By: alexchow

Differential Revision: D5951272

fbshipit-source-id: 16858457fe61a2f0813743a13ee54d69f83795a9