Fix a bug unwinding from the onFunctionExit hook
Summary:
onFunctionExit runs after the frame has been torn down, and can throw
from a user profiler.

When it does, we can't run any fault funclets that appear to be
required.

Prior to this diff, the new test case asserts in dbg builds, but runs the catch block in
opt builds (despite the frame being torn down), and in repo-auth mode
issues a fatal error because hhbbc (correctly) thought the catch block
was unreachable.

Reviewed By: jano

Differential Revision: D5954923

fbshipit-source-id: f74f5fb317c941dbe602df153e5f4bda5f8be242