Only support jemalloc 5 when using 1G huge pages
Summary:
We plan to have tighter integration with jemalloc 5 (and also improve jemalloc 5 to better support our use case).
This will drop support for using jemalloc 4 to manage the huge pages, which we don't think anyone is doing at the moment.
OSS users of HHVM will still be able to build with jemalloc 4, just without the 1G huge page features.

Reviewed By: interwq

Differential Revision: D5957188

fbshipit-source-id: 4fd7bd5d8517c188398d597bd5ce5c32f7b7dce7