pattern for closure creation in argument push list
Summary:
We generally give up on seeing FPassL _n as we can't reliably tell if the param is passed by reference.
However, there's a special pattern for the stylized safe use of this instruction in compiling pipes. This diff
extends that pattern to cover the case in which the other arguments include a closure that's created inline,
rather than just being FPass'd

Reviewed By: hubyrod

Differential Revision: D5963439

fbshipit-source-id: 35a688754b918b76daa62ba2510ee9741493bf03