Fix semdiff reporting failure in exception-heavy code
Summary: In code that can throw a number of exceptions, there can be multiple paths to the same point. For a certain test, semdiff was failing while processing IContFlow Unwind instructions even though the code is semantically correct. This was due to the previous_assumptions bound. Increasing this bound (5 was chosen arbitrarily) causes semdiff to work correctly.

Reviewed By: nbenton

Differential Revision: D5963353

fbshipit-source-id: 8b3f070e7c8ac6142932a4d0763892009722b924