aarch64 - correction for LSE Atomics
Summary:
Running a MOP on a larger lab machine exposed unexpected failures when -vEval.JitPGO=1 was used.  The issue was traced to the use of the LSE Atomic instructions.  The implementation of
PR#7857 was not complete.  The decqmlock Vinstr was expected to set the PSW.  Without this
change arbitrary PSW values were returned when the hardware supported the LSE instructions.

The hphp/test/quick/setg-dtor.php test reliably exposed the problem.
Closes https://github.com/facebook/hhvm/pull/7997

Differential Revision: D5855191

Pulled By: mxw

fbshipit-source-id: c5dd9fb3dfa1f10f6e5e80ca2b2d5719ef287ca7