Preserve break before first arrow in MemberSelection expressions
Summary:
The first MemberSelection in a chained MemberSelection expression sometimes has special meaning. For instance, it might construct a builder or a query object, and all subsequent chained calls are builder methods or query filters. Keeping the first call in the chain joined with the receiver calls it out as different from the rest of the calls and helps with readability.

Since hackfmt does not have access to semantic information (i.e., whether the first method invocation in a chain of calls belongs to a different type than the remainder), we just add a newline before the first arrow if one was present in the original source.

Reviewed By: arxanas

Differential Revision: D5967011

fbshipit-source-id: 46c389cab6f5316573eb2aa8c0946e3cdd391561