Make assignment operator less sticky
Summary: This change reduces the cost of the split following an assignment operator from 1 to 0. I don't think splitting here is inherently undesirable, other than that it sometimes increases the indentation level of the right-hand-side. Increased indentation is already penalized in the line splitter when it leads to more overflow, so I don't think further penalizing splitting here by having a cost is necessary.

Reviewed By: arxanas

Differential Revision: D5967115

fbshipit-source-id: 184db21216a252fb4b3f20fb4e51fa20cec2b9e9