Add support for 'using' feature
Summary:
Conceptually, this is just a syntax transformation into try/finally. I
put most of the work in the emitter, however, for two reasons: it needs unnamed
locals, and the full-function form felt like it could get messy to do on the
AST.

The main change is the addition of a stack of PendingFinally structs, which
store all the information needed to emit a finally block. For normal
try/finally blocks, elements of this stack don't live very long, but for
whole-function using statements, they stick around until the end of the
function. I cleaned up a few things along the way that were either blocking me
or just making me twitch:
- m_memoizedFunclets is now keyed by ConstructPtr rather than StatementPtr,
  because one using statement can generate an arbitrary number of finally
  funclets (but there's one Expression corresponding to each one).
- newFaultRegionAndFunclet() was unused and has been deleted.
- The code to emit a TryStatement has been split into emitTryCatch() and
  emitPendingFinally(), since these two pieces can happen far apart from each
  other for a using statement.
- EmitterVisitor::createRegion() now prefers to take a LabelScopePtr, since
  that's all it needs from the StatementPtr it used to take.
- FinallyStatement has been deleted. As far as I can tell, it only existed to
  hold on to the LabelScope for the finally block, and act as a key in
  m_memoizedFunclets. But we can accomplish the same thing by just ensuring
  there's always a StatementList in the TryStatement, even when the finally
  block is empty.
- bytecode.specification was missing documentation for its ObjMethodOp
  immediate.
- FuncEmitter::freeUnnamedLocal() assumed that callers were freeing the most
  recently allocated unnamed local without verifying or documenting that
  assumption.
- The parse error for goto into a finally block used "loop or switch
  statement". I fixed it to say "finally statement", and added a case for goto
  into/across a using statement.

Reviewed By: paulbiss

Differential Revision: D5916508

fbshipit-source-id: 595cccb90b74c47691ab6f53f4871458a25337d2