Explicitly comparing `Pos.t`s in `hh_single_parse`
Summary:
Using the Full Fidelity Parser as a front-end for the type checker produces differences compared to `Parser_hack`. One specific category of differences is off-by-one errors in `Pos.t`s.

Using `hh_single_parse` is the easiest way to debug differences, but it didn't yet compare the resulting `Pos.t`-things, because it used to compare S-expressions (in which all `Pos.t`-things are rendered simply as `p`). Now it does.

Reviewed By: ericlippert, vladima

Differential Revision: D5913831

fbshipit-source-id: f038f4a6de572460fafa3e61c27afedffe78313a