guard literal equality of instructions
Summary:
There are several checks for instruction equality that lead to accepting equivalence. In most cases, these are sound because the instructions that use locals are already handled. This is a brittle approach, particularly in case new instructions are added. Futhermore, instruction `NewLikeArrayL` is not handled correctly because of this.

This diff guards each instruction equality check (that effects equivalence) with a whitelist of instructions that do not access locals, and also adds a rule for `NewLikeArrayL`. Futhermore, `| _, _ -> None` patterns on instructions are replaced with explicit patterns to ensure that all instructions are handled (otherwise, it is hard to know that adding a new instruction has broken semdiff).

Reviewed By: andrewjkennedy

Differential Revision: D5924481

fbshipit-source-id: 8bf7689588f5ec19958cd9d4e863d9c9996eea68