audit SetWithRefRML rule
Summary: I've verified that the behavior of `SetWithRefRML` reads from the local (and does not write to it); the current handling appears sound. Thus I've removed the comment "I'm guessing a bit here!". However, the instruction should not be generated by HackC or HHVM (according to some HHVM people I talked to, anyhow), so I've changed to behavior to simply reject the instruction.

Reviewed By: andrewjkennedy

Differential Revision: D5942082

fbshipit-source-id: 5e55ecb181c1456cbf1cb75b6bc6774bcdd2b209