Fix invoking magic calls with splats
Summary:
If you perform a magic call with the splat operator, we were pulling
the non-splatted arguments off the stackk in the wrong order, meaning they would
be presented to __call or __callStatic in reverse order. Fix this by copying
values off of the stack in reverse order, meaning they're put into the
"arguments" array in the proper order (which we were already doing properly
elsewhere).

Reviewed By: markw65

Differential Revision: D5979598

fbshipit-source-id: b50b9d4565e7e06897f5b4c50a79d1a9ec8cea73