Address flakey lsp integration test
Summary:
I got reports that the lsp integration tests are consistently failing. Inside `test_hover`, it calls `prepare_environment`, which calls `self.check_cmd(...)`, which attempts to launch `hh check`. But it's getting back this on stderr:

   Error: Ran out of retries, giving up\!

I figure the job of the lsp integration tests is to test lsp. If we can't spin up hh, well, that shouldn't report as a failure of LSP.

This diff detects that situation and skips the test.

Differential Revision: D5984275

fbshipit-source-id: d6f71382a68a681fc7a1c84f01c24a6c47f83eaa