Another cause of flakey LSP tests
Summary:
Each LSP integration tests attempts to start hh_server, wait until it's responsive as indicated by "hh check" returning "No errors\!" within 20 one-second retries, then allows 500ms to establish a persistent connection with it, and then enters in dialog over that persistent connection.

The previous diff addressed cases (observed in sandcastle but not in testing) where hh_server simply failed to start, by simply skipping the test in this case. I believe that hh_server might fail to start in cases where the machine is overloaded or out of memory.

The current diff addresses cases where the machine is so busy that hh_server isn't able to establish a persistent connection within 500ms. In these cases we'll get back the message "Server busy". Again, we simply skip over the test.

We might consider changing it to do retries. But that will only delay the question -- what if it's still busy after 10 retries? after 20? after 2 days? The only good answer is to skip the test if hh_server isn't responsive enough.

(We don't generally observe hh_server being unresponsive. If it is, I trust it will be caught fine in other situations).

Reviewed By: oulgen

Differential Revision: D5988628

fbshipit-source-id: 899312a5d0a731cb3525461898c09773e76e64b8