Don't box object properties obtained via read APIs.
Summary: We have 2 code paths that want to read object properties as zvals; 'read' as zval means boxing properties that aren't boxed, so they're really writes. Stop abusing read APIs for this; add a mode to o_realProp that boxes if needed and convert these code paths to use it. Since the boxing now happens in one place, we can easily forbid it for immutable properties later.

Reviewed By: markw65

Differential Revision: D5985028

fbshipit-source-id: 8738e6aaf78d60bcd983d4792f8f0c9e82fad2c6