Handle alias_class in hhbbc
Summary:
Calls to alias_class in namespaces don't get converted to the AliasCls
bytecode; if hhbbc detects that an FPushFuncU really is a call to
alias_class, it needs to update the index. Unfortunately, we can't do
that incrementally, because we might already have taken advantage of
some fact thats no longer true (exact resolution for a class thats no
longer unique, for example). So we throw an exception and start over.

Reviewed By: ricklavoie

Differential Revision: D5986157

fbshipit-source-id: 1e734a6b114ebf69e2c9782c568ac7571acfa69d