Don't resolve namespace fallback funcs in hphpc
Summary:
This is the last remaining thing that requires knowledge of all the
declared functions in hphpc.

hhbbc handles almost everything. If assert was called from within a
namespace, it will still be resolved to a builtin, but it won't get
the ini_get guard, which means that side effects of arguments will be
observed; but we already have that bug in sandbox mode.

Reviewed By: ricklavoie

Differential Revision: D5975243

fbshipit-source-id: 6986616114936d769be327a3f5aedd5c3af2640e