dismiss diagnostics if hh_server goes down
Summary:
When hh_server goes down, we should flush all diagnostics that we've reported so far. (If hh_server comes back again, then we can restore them).

Reason: imagine if we were in a state with errors shown in Nuclide, and then did a rebase (say) or some other operation that killed hack and later we fixed the errors while Hack was busy restarting. The new Hack instance would have no way of knowing which errors were already shown in Nuclide, and so would have no way of dismissing them. The only solution is to dismiss them as soon as hh_server goes down.

Reviewed By: arxanas

Differential Revision: D6004209

fbshipit-source-id: 6465ac62be2310aee8f3096355d93593ca807d86