more informative failures of lsp integration tests
Summary:
When an LSP integration test fails with the wrong *content* of messages, it helpfully displays a diff. That lets you examine the test-failure-report and see precisely what went wrong.

But when an LSP integration test fails with the wrong *number* of messages, it doesn't bother with a diff, but merely displays something like "expected 5 but got 6 messages". So you can't learn anything from the test-failure-report.

This change makes it so it does provide useful information in the latter failure case.

Differential Revision: D6006998

fbshipit-source-id: 8839dc6331f4432a8a3ef3259dfdef4f08d98867