Fix This type annotation to not behave like self.
Summary:
D4892972 implements the this typehint to be checked like self.
This diff corrects it to perform the proper typecheck using the context class.
The options available are:
```
Eval.ThisTypeHintLevel = 0  // <- Typecheck like mixed.
Eval.ThisTypeHintLevel = 1  // <- Typecheck like self.
Eval.ThisTypeHintLevel = 2  // <- Typecheck like this, but soft failure even for hard typehints.
Eval.ThisTypeHintLevel = 3  // <- Typecheck like this.

Eval.CheckThisTypeHints   // <- continues to exist but now enables/disables
                          //    HHBBC optimization that is only compatible
                          //    with Eval.ThisTypeHintLevel=3.
```

The changes to allow multiple behaviors (notably `Eval.ThisTypeHintLevel=2`) are a result of conversations with
KendallHopkins about easing the transition to hard `this`.

Depends on D4892972

Reviewed By: markw65

Differential Revision: D5341430

fbshipit-source-id: 78599030423169c93c8f8399d3afa7ef9b44fc2d