Informant uses most significant report when churning queue; add test
Summary:
There was a bug in the Informant which is caught by this new unit test.

The bug arose because we would us the greatest "significance" of all the
"ready" changes in the transition queue, but when deciding on the actual
report to form, we were only looking at the last transition in the queue.

Here's why that doesn't work.
Suppose the repo starts at Rev 1.

Then Watchman notifies us of 3 transitions:
Merge_base_changed to Rev 200,
State_enter to rev 230
State_leave at 230

Suppose that querying hg for the revision number for 200 takes 3 seconds
(i.e. 3 server ticks).

This results in the transition queue holding all 3 of these events to be
processed. After the 3 seconds, when we churn the queue we handle all
3 of these transitions and discover that we have moved >100 revisions
away. I.e., we've made a "significant" transition.

When deciding what action to take, we peek only at the last
transition in the queue, we decide that a "significant" move with a
State_leave should be a "Move_along". But actually we should
Restart_server due to the Merge_base_changed.

This diff adds a test for this, and fixes the behavior.

In a prior world, we were restarting on State_leave transitions.
Now, we restart on Changed_merge_base transitions.

The behavior made sense in the old world - whenever we see
any sequence of repo transitions that contains a "significant" jump then
we peek at the most recent transition (the last in the queue) and either
(1) The last transition is a State_leave so we should restart the
server (and hopefully we end up loading a better saved state for the new
revision we settled at)
or
(2) The last transition is a State_enter, so we shouldn't restart the
server yet until the State_leave arrives.

In the new world with Changed_merge_base, this is not how we
should do things anymore.

Reviewed By: jamesjwu

Differential Revision: D5960781

fbshipit-source-id: 447271fefb6c0703bfae25effe16a684bc0958fd