Clean up code for checking default parameter errors
Summary:
There was some code in the type checker trying to do three things all at the same time; this was hard for me to reason about so I've split it up into three methods, each of which does one thing well.

* Map parameter declarations to their types
* Count the minimum number of required arguments to a call
* Give an error if a no-default-value parameter comes after a default-value parameter.

Note that the latter need not be checked by the type checker at all; we can check it purely based on the syntax, after the initial parsing pass.

I've added this check to the full-fidelity-parser post-parse error pass; when we switch over to the FFP we can remove the check from the typing pass.

 

Differential Revision: D5943333

fbshipit-source-id: a417d6b1811de0593374f1fc51bf003963f05bd3