Fix some trait related constructor issues
Summary:
A same-named function is a constructor if its defined by the class, or
by one of its used traits. The runtime called it a constructor if it
was defined by the class or a trait anywhere in the hierarchy.

If the class defines a same-named function, and a used trait defines
an __construct, or a used trait defines a same-named function and the
class defines an __construct, or a trait defines both, its an error;
again, the runtime was checking for traits used anywhere in the
hierarchy.

We can fix both these issues by checking that func->cls() matches the
current Class - that will be the case if and only if the func came
from the current pre-class, or one of its used traits.

In the new test case, we thought X::x was a constructor, and
incorrectly raised an error for both Y and Z.

Reviewed By: ricklavoie

Differential Revision: D6051286

fbshipit-source-id: fa59ff665bbcbe8d4fae329c4b13b7c318c67099