Fix some type system bugs
Summary:
I hit an assert compiling something like:

  $x = null;
  if (condition) $x = new X;
  $x?->foo();
  return $x;

The issue was that on the first pass, we couldn't resolve condition,
so $x had type ?X, but on the second pass, we knew it was false, so $x
had type Null.

However, FPushMethodD will attempt to "refine" the type of the local
(if its called on one). Since $x isn't a specialized object type, it
refines it to OptObj, making it worse than the already known Null.

Then we return that, and OptObj isn't a subtype of OptObj=X.

To fix that, I modified refineLocation to take the intersection of the
two types, falling back to the current behavior in the case that the
intersection isn't more refined. Now Null stays as Null, and Null *is*
a subtype of OptObj=X.

But that resulted in a new problem where locals were being given type
TBottom, which turned out to be because couldBe is too lenient when
comparing specialized object types; if either was DObj::Sub, it
behaved as if both were. This meant there were cases where couldBe
returned true, but intersection_of (correctly) returned TBottom. This
was affecting InstanceofD, because we first use subtypeOf and !couldBe
to check for a known result, and then assume that refineLocation can't
produce TBottom for the remaining cases.

Reviewed By: ricklavoie

Differential Revision: D6057257

fbshipit-source-id: 56a85a135b33acc25b0df86fc7661b3608130c9b