Hack AST and parsing for "using" construct
Summary:
This is the first stage of Hack support for the `using` construct.
For now, we implement
- A new constructor `Using` in both the AST and NAST
- Appropriate visitor boilerplate
- Legacy parsing support for the block-scoped version of the construct
- Minimal checks in the Naming phase

This means that in non-strict mode we can type-check one of the tests from the HHVM suite, which is
copied into new directory `hphp/hack/tests/typecheck/using`. Once full type-checker support has been
added, this will be expanded into a properly type-safe test.

Reviewed By: oulgen

Differential Revision: D5920624

fbshipit-source-id: 02b910509bf5065b844fda60c2d7f27f7049dbaa