Refactor choosing init approach; log name of approach
Summary:
The bulk of this diff is a refactor of how we choose the init
approach. It is a simplification and increases readibility.

I also add another return value of a string name for the approach attempted
which was completely missing before. i.e., we conflated a lot
of stuff under the "fresh" init type (the desired approach failed or
some args or configs were toggled) and thus the data wasn't able
to show us why we have a "fresh" init type. So we have two strings now -
the approach_name is what we wanted Server_init to try to do, and the
init_type which is what Server_init actually ended up doing (for example, we might
ask Server_init to try a mini state load, but it might fail and then end up
having init_type = "fresh")

And then I log that approach name

Reviewed By: jamesjwu

Differential Revision: D6047825

fbshipit-source-id: 09ecaf514fe549b7115a1b03bae752b21eee7353