Add a file hash to the FileInfo object
Summary:
While looking through the data for incremental init, I realized I'm not seeing the results that I want because the overhead for incremental mode is just very very high: In order to take advantage of the old version of files, we need to do a whole lot of work to declare, redeclare, and compare them. This added overhead meant that initialization times were worse for p50, even though they are better for p99.

I decided that instead of all of this, a simpler approach would work better.

This first diff introduces the Decl_Hash module, and adds it to our FileInfo object.

1. For each file, we save a hash of the file's AST in the FileInfo object. We make sure that we store the *decl* ASTs, without method bodies, so that two files with modified method bodies hash to the same thing.
2. This FileInfo object is stored in our saved state already. We then can compare a file's old saved hash with the new hash to see if they are equal. If they are, we can avoid typecheck all the file's dependencies, and instead only typecheck the file itself.
3. The association of a file with its hash can also be extended to be used in incremental mode.

Reviewed By: alexchow

Differential Revision: D5972887

fbshipit-source-id: 8258d4f7c6eed15ce0827886d4a27599af4c7453